+++ src/geometry/transform.rs
@@ -3,57 +3,64 @@
  * Copyright 2025 Sira Pornsiriprasert <code@psira.me>
  */
 
-//! Transform object position and orientation. Provide [Transform] trait.
+//! Provides the [`Transform`] trait and macros for implementing it.
 
 use nalgebra::{Point3, Translation3, UnitQuaternion};
 
-/// Transform object in 3D Cartesian CS.
+/// Trait for transforming objects in 3D Cartesian CS.
 pub trait Transform {
+    /// Get the object position.
     fn position(&self) -> Point3<f64>;
+    /// Get the object orientation.
     fn orientation(&self) -> UnitQuaternion<f64>;
+    /// Set the object position.
     fn set_position(&mut self, position: Point3<f64>);
+    /// Set the object orientation.
     fn set_orientation(&mut self, orientation: UnitQuaternion<f64>);
+    /// Translate the object.
     fn translate(&mut self, translation: &Translation3<f64>);
+    /// Rotate the object.
     fn rotate(&mut self, rotation: &UnitQuaternion<f64>);
+    /// Rotate the object using the anchor point as the center of rotation.
     fn rotate_anchor(&mut self, rotation: &UnitQuaternion<f64>, anchor: &Point3<f64>);
 }
 
-/// Implement shallow [Transform] for objects with no children.
+/// Macro to implement shallow [`Transform`] for types with no children.
 #[macro_export]
 macro_rules! impl_transform {
     ($type:ty) => {
         impl Transform for $type {
-            /// Get the object position.
+            #[inline]
             fn position(&self) -> Point3<f64> {
                 self.position
             }
 
-            /// Get the object orientation.
+            #[inline]
             fn orientation(&self) -> UnitQuaternion<f64> {
                 self.orientation
             }
 
-            /// Set the object position.
+            #[inline]
             fn set_position(&mut self, position: Point3<f64>) {
                 self.position = position;
             }
 
-            /// Set the object orientation.
+            #[inline]
             fn set_orientation(&mut self, orientation: UnitQuaternion<f64>) {
                 self.orientation = orientation;
             }
 
-            /// Translate the object.
+            #[inline]
             fn translate(&mut self, translation: &Translation3<f64>) {
                 self.position = translation.transform_point(&self.position);
             }
 
-            /// Rotate the object.
+            #[inline]
             fn rotate(&mut self, rotation: &UnitQuaternion<f64>) {
                 self.orientation = rotation * &self.orientation;
             }
 
-            /// Rotate the object using the anchor point as the center of rotation.
+            #[inline]
             fn rotate_anchor(&mut self, rotation: &UnitQuaternion<f64>, anchor: &Point3<f64>) {
                 let local_position = self.position - anchor;
                 self.position =