+++ src/special/mod.rs
@@ -1,9 +1,0 @@
-/*
- * Magba is licensed under The 3-Clause BSD, see LICENSE.
- * Copyright 2025 Sira Pornsiriprasert <code@psira.me>
- */
-
-//! Special mathematical functions.
-
-mod cel;
-pub use cel::*;