+++ src/util.rs
@@ -5,6 +5,4 @@
 
 //! Utilities for Magba
 
-pub fn is_close(a: f64, b: f64, rtol: f64) -> bool {
-    (a - b).abs() <= rtol * b.abs()
-}
+pub use crate::crate_util::assert_close_vector_elem;